Zod: fix handling of functions that return nothing
Since `undefined` is not a valid Convex value, functions that return nothing have a return value of `null` (not `undefined`). The Zod types are already wired correctly (a return validator of `z.null()` is accepted in this case), but we pass `undefined` as a value to Zod, which leads to Zod throwing an error at runtime. This pull request fixes this issue.